# frontend/openapi/wst-openapi.yaml
openapi: 3.1.0
info:
  title: Workshop Management & Student Practical Training (WST) API
  version: 1.0.0
  summary: Contract-first REST API for the WST academic MVP.
  description: |
    Source of truth for the PostgreSQL schema, Node.js backend, automated tests, React client and documentation.
    Requirement prefix: WST. Every operation carries `x-requirement-id` and `x-priority`.

    ## Conventions
    - Base path `/api/v1`. JSON bodies (multipart only for attachment upload). Bearer JWT. Operations with `security: []` are public.
    - Identifiers are UUIDs. Human-readable numbers (`jobNumber`, `poNumber`, `invoiceNumber`, `certificateNumber`) are server-generated, unique, immutable, and never localized or reordered by the backend.
    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD`.
    - Money is `{amount, currency}`: `amount` is a decimal string (never a float) and `currency` an ISO-4217 code. One system currency is configured in finance settings; documents in another currency are rejected with 422. Part quantities are whole numbers. Rates and metric values are decimal strings.
    - Optional properties without a value are omitted, never null. `readOnly` properties are server-controlled. Request schemas forbid unknown properties, so clients cannot submit ids, audit fields, stock balances, invoice totals, or permissions.
...

    ## Collections
...

    ## Authorization
...

    ## Concurrency and idempotency
...

    ## Localization
...

    ## Errors and rate limiting
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /training-terms:
    get:
      tags: [Training]
      operationId: listTrainingTerms
      summary: List training terms
      description: "Requires `training.read`. Sortable: startDate, name (default -startDate)."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [training.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - { name: status, in: query, schema: { type: string, enum: [PLANNED, ACTIVE, CLOSED] } }
      responses:
        '200': { description: Terms, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingTermPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Training]
      operationId: createTrainingTerm
      summary: Create a training term
      description: "Requires `training.manage`."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [training.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingTermCreateRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingTerm' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /training-terms/{termId}:
    parameters:
      - $ref: '#/components/parameters/TermId'
    patch:
      tags: [Training]
      operationId: updateTrainingTerm
      summary: Update a training term
      description: "Requires `training.manage`. CLOSED terms are read-only (409 INVALID_STATE_TRANSITION)."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [training.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingTermUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingTerm' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /courses:
    get:
      tags: [Training]
      operationId: listCourses
      summary: List courses
      description: "Requires `training.read` or `students.self` (students see only courses they are enrolled in). Sortable: code, createdAt."
      x-requirement-id: [WST-FR-10, WST-FR-01]
      x-priority: Must
      x-permissions: [training.read, students.self]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
        - { name: termId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: status, in: query, schema: { type: string, enum: [DRAFT, ACTIVE, ARCHIVED] } }
      responses:
        '200': { description: Courses, content: { application/json: { schema: { $ref: '#/components/schemas/CoursePage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Training]
      operationId: createCourse
      summary: Create a course with its required tasks and completion rules
      description: "Requires `training.manage`. Course tasks reference the practical task library and flag which are required for completion."
      x-requirement-id: [WST-FR-10, WST-FR-12]
      x-priority: Must
      x-permissions: [training.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CourseCreateRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Course' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /courses/{courseId}:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    patch:
      tags: [Training]
      operationId: updateCourse
      summary: Update a course
      description: "Requires `training.manage`. Changing required tasks or completion rules of an ACTIVE course that has enrollments returns 409 RESOURCE_IN_USE so eligibility already granted is not silently invalidated."
      x-requirement-id: [WST-FR-10, WST-FR-12]
      x-priority: Must
      x-permissions: [training.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CourseUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Course' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /students:
    get:
      tags: [Training]
      operationId: listStudents
      summary: List students
      description: "Requires `training.read` (mentors see only students of their sessions). Students cannot list other students (403). Sortable: studentNumber, displayName."
      x-requirement-id: [WST-FR-10, WST-FR-01]
      x-priority: Must
      x-permissions: [training.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
        - { name: status, in: query, schema: { type: string, enum: [ACTIVE, INACTIVE] } }
        - { name: groupId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: courseId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
      responses:
        '200': { description: Students, content: { application/json: { schema: { $ref: '#/components/schemas/StudentPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Training]
      operationId: createStudent
      summary: Create a student profile for a STUDENT user
      description: "Requires `training.manage`. The referenced user must hold the STUDENT role. The profile stores no contact data beyond the user account."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [training.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StudentCreateRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Student' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /students/{studentId}:
    parameters:
      - $ref: '#/components/parameters/StudentId'
    get:
      tags: [Training]
      operationId: getStudent
      summary: Get a student profile
      description: "Requires `training.read` or `students.self`. A student may read only their own profile; any other studentId returns 404 without confirming existence."
      x-requirement-id: [WST-FR-01, WST-FR-10]
      x-priority: Must
      x-permissions: [training.read, students.self]
      responses:
        '200': { description: Student, content: { application/json: { schema: { $ref: '#/components/schemas/Student' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Training]
      operationId: updateStudent
      summary: Update a student profile
      description: "Requires `training.manage`."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [training.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StudentUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Student' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /training-groups:
    get:
      tags: [Training]
      operationId: listTrainingGroups
      summary: List training groups
      description: "Requires `training.read`. Sortable: name, createdAt."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [training.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - { name: courseId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: status, in: query, schema: { type: string, enum: [ACTIVE, CLOSED] } }
      responses:
        '200': { description: Groups, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingGroupPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Training]
      operationId: createTrainingGroup
      summary: Create a training group for a course
      description: "Requires `training.manage`."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [training.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingGroupCreateRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingGroup' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /training-groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    patch:
      tags: [Training]
      operationId: updateTrainingGroup
      summary: Update or close a training group
      description: "Requires `training.manage`."
      x-requirement-id: [WST-FR-10]
      x-priority: Must
      x-permissions: [training.manage]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingGroupUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/TrainingGroup' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token issued by login or refresh. The token identifies the user only; roles, permissions and scopes are resolved server-side on each request."
  parameters:
    Page: { name: page, in: query, description: "1-based page number.", schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, description: "Items per page, maximum 100.", schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Sort: { name: sort, in: query, description: "Comma-separated fields; prefix - for descending. Only fields named by the operation are accepted.", schema: { type: string, maxLength: 100, pattern: '^-?[A-Za-z][A-Za-z0-9]*(,-?[A-Za-z][A-Za-z0-9]*)*$' } }
...
    Query: { name: q, in: query, description: "Case-insensitive text search.", schema: { type: string, minLength: 1, maxLength: 100 } }
...
    TermId: { name: termId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
    CourseId: { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
    StudentId: { name: studentId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
    GroupId: { name: groupId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
  responses:
    BadRequest:
      description: "Malformed request, unknown filter or sort field, or invalid query parameter."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: "Missing, invalid or expired credentials."
      headers: { WWW-Authenticate: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: "Authenticated but lacking the required permission. Rows outside the caller's scope are reported as 404 instead."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: "Resource does not exist or is outside the caller's row scope."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: "Invalid state transition, missing customer approval, insufficient stock, concurrency conflict, separation-of-duties violation or duplicate business operation. See Error.code."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ValidationError:
      description: "Semantic validation failed; details identify each offending field."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time, description: "RFC 3339 UTC timestamp." }
    DateOnly: { type: string, format: date }
...
    LocalizedName:
      type: object
      required: [en]
      properties: { en: { type: string, minLength: 1, maxLength: 200 }, ar: { type: string, minLength: 1, maxLength: 200 } }
    RecordMeta:
      type: object
      required: [id, createdAt, updatedAt, createdBy, updatedBy]
      properties:
        id: { $ref: '#/components/schemas/Uuid', readOnly: true }
        createdAt: { $ref: '#/components/schemas/Timestamp', readOnly: true }
        updatedAt: { $ref: '#/components/schemas/Timestamp', readOnly: true }
        createdBy: { $ref: '#/components/schemas/Uuid', readOnly: true }
        updatedBy: { $ref: '#/components/schemas/Uuid', readOnly: true }
    PageInfo:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties: { page: { type: integer, minimum: 1 }, pageSize: { type: integer, minimum: 1, maximum: 100 }, totalItems: { type: integer, minimum: 0 }, totalPages: { type: integer, minimum: 0 } }
...
    ErrorCode:
      type: string
      enum: [BAD_REQUEST, UNAUTHENTICATED, INVALID_CREDENTIALS, TOKEN_EXPIRED, FORBIDDEN, NOT_FOUND, VALIDATION_FAILED, RATE_LIMITED, INTERNAL_ERROR, SERVICE_UNAVAILABLE, INVALID_STATE_TRANSITION, VERSION_CONFLICT, DUPLICATE_RESOURCE, IDEMPOTENCY_CONFLICT, RESOURCE_IN_USE, CUSTOMER_APPROVAL_REQUIRED, JOB_STAGE_NOT_ALLOWED, JOB_ASSIGNMENT_REQUIRED, CHECKLIST_INCOMPLETE, QUALITY_CHECK_REQUIRED, INVOICE_REQUIRED, INSUFFICIENT_STOCK, REVERSAL_EXCEEDS_ISSUED, RESERVATION_INVALID, SCHEDULE_CONFLICT, CONFLICT_NOT_OVERRIDABLE, SEPARATION_OF_DUTIES_VIOLATION, DUPLICATE_APPROVAL, RECEIPT_EXCEEDS_ORDERED, PAYMENT_AMOUNT_MISMATCH, ASSESSMENT_LOCKED, CERTIFICATE_NOT_ELIGIBLE, ATTACHMENT_INVALID, ATTACHMENT_NOT_LINKABLE, PAYLOAD_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE, EXPORT_NOT_READY, EXPORT_EXPIRED]
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        field: { type: string, description: "JSON Pointer to the offending request field." }
        code: { type: string, description: "Machine-readable detail code, for example REQUIRED, TOO_LONG, MILEAGE_LOWER_THAN_RECORDED." }
        message: { type: string }
        params: { type: object, additionalProperties: { type: string }, description: "Named values for message templates, for example available and requested quantities." }
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string, description: "Human-readable message in the negotiated language." }
        details: { type: array, items: { $ref: '#/components/schemas/ErrorDetail' } }
        requestId: { type: string, description: "Correlation id also present in logs and audit events." }
...
    TrainingTerm:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [organizationScopeId, name, startDate, endDate, status]
          properties: { organizationScopeId: { $ref: '#/components/schemas/Uuid' }, name: { type: string, maxLength: 120 }, startDate: { $ref: '#/components/schemas/DateOnly' }, endDate: { $ref: '#/components/schemas/DateOnly' }, status: { type: string, enum: [PLANNED, ACTIVE, CLOSED] } }
    TrainingTermCreateRequest:
      type: object
      required: [organizationScopeId, name, startDate, endDate]
      additionalProperties: false
      properties: { organizationScopeId: { $ref: '#/components/schemas/Uuid' }, name: { type: string, minLength: 1, maxLength: 120 }, startDate: { $ref: '#/components/schemas/DateOnly' }, endDate: { $ref: '#/components/schemas/DateOnly' } }
    TrainingTermUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties: { name: { type: string, minLength: 1, maxLength: 120 }, startDate: { $ref: '#/components/schemas/DateOnly' }, endDate: { $ref: '#/components/schemas/DateOnly' }, status: { type: string, enum: [PLANNED, ACTIVE, CLOSED] } }
...
    CourseTask:
      type: object
      required: [taskId, required]
      additionalProperties: false
      properties: { taskId: { $ref: '#/components/schemas/Uuid' }, required: { type: boolean } }
    Course:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [organizationScopeId, code, name, termId, tasks, minimumAttendancePercent, status]
          properties: { organizationScopeId: { $ref: '#/components/schemas/Uuid' }, code: { type: string, minLength: 1, maxLength: 30 }, name: { $ref: '#/components/schemas/LocalizedName' }, termId: { $ref: '#/components/schemas/Uuid' }, description: { type: string, maxLength: 2000 }, tasks: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/CourseTask' } }, minimumAttendancePercent: { type: integer, minimum: 0, maximum: 100 }, status: { type: string, enum: [DRAFT, ACTIVE, ARCHIVED] } }
    CourseCreateRequest:
      type: object
      required: [organizationScopeId, code, name, termId, tasks, minimumAttendancePercent]
      additionalProperties: false
      properties: { organizationScopeId: { $ref: '#/components/schemas/Uuid' }, code: { type: string, minLength: 1, maxLength: 30 }, name: { $ref: '#/components/schemas/LocalizedName' }, termId: { $ref: '#/components/schemas/Uuid' }, description: { type: string, maxLength: 2000 }, tasks: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/CourseTask' } }, minimumAttendancePercent: { type: integer, minimum: 0, maximum: 100 } }
    CourseUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties: { name: { $ref: '#/components/schemas/LocalizedName' }, description: { type: string, maxLength: 2000 }, tasks: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/CourseTask' } }, minimumAttendancePercent: { type: integer, minimum: 0, maximum: 100 }, status: { type: string, enum: [DRAFT, ACTIVE, ARCHIVED] } }
    Student:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [userId, studentNumber, displayName, status]
          properties: { userId: { $ref: '#/components/schemas/Uuid' }, studentNumber: { type: string, minLength: 1, maxLength: 30 }, displayName: { type: string }, status: { type: string, enum: [ACTIVE, INACTIVE] } }
    StudentCreateRequest:
      type: object
      required: [userId, studentNumber]
      additionalProperties: false
      properties: { userId: { $ref: '#/components/schemas/Uuid' }, studentNumber: { type: string, minLength: 1, maxLength: 30 } }
    StudentUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties: { studentNumber: { type: string, minLength: 1, maxLength: 30 }, status: { type: string, enum: [ACTIVE, INACTIVE] } }
    TrainingGroup:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [name, courseId, status, enrolledCount]
          properties: { name: { type: string, maxLength: 120 }, courseId: { $ref: '#/components/schemas/Uuid' }, status: { type: string, enum: [ACTIVE, CLOSED] }, enrolledCount: { type: integer, minimum: 0 } }
    TrainingGroupCreateRequest:
      type: object
      required: [name, courseId]
      additionalProperties: false
      properties: { name: { type: string, minLength: 1, maxLength: 120 }, courseId: { $ref: '#/components/schemas/Uuid' } }
    TrainingGroupUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties: { name: { type: string, minLength: 1, maxLength: 120 }, status: { type: string, enum: [ACTIVE, CLOSED] } }
...
    TrainingTermPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/TrainingTerm' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
    CoursePage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/Course' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
...
    StudentPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/Student' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
    TrainingGroupPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/TrainingGroup' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
...